Extract out Builder supertype
