+++ solr/solrj/src/java/org/apache/solr/client/solrj/impl/LBHttp2SolrClient.java
@@ -34,4 +34,16 @@ public class LBHttp2SolrClient extends LBHttpSolrClientBase {
     public Set<String> getUrlParamNames() {
         return solrClient.getUrlParamNames();
     }
+
+    public static class Builder extends LBHttpSolrClientBuilderBase<LBHttp2SolrClient, LBHttp2SolrClient.Builder, Http2SolrClient> {
+
+        public Builder(Http2SolrClient http2Client, Endpoint... endpoints) {
+            super(http2Client, endpoints);
+        }
+
+        @Override
+        public LBHttp2SolrClient build() {
+            return new LBHttp2SolrClient(this);
+        }
+    }
 }
